requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

# Google Cloud & BigQuery
google-cloud-bigquery>=3.10.0
google-cloud-storage>=2.16.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# AI Processing
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.18.0

# Database Integration
supabase>=2.0.0
# Use the correct PyPI package name and available versions
postgrest>=0.10.6
websockets>=13.0

# Airtable Integration
pyairtable>=2.1.0

# Configuration & Data Formats
PyYAML>=6.0
python-dotenv>=1.0.0

# File Processing & I/O
pathlib2>=2.3.7
chardet>=5.1.0

# Logging & Monitoring
colorlog>=6.7.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Optional: Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Security
cryptography>=41.0.0
requests>=2.31.0

# Performance
psutil>=5.9.0
tqdm>=4.65.0

# Data Validation
pydantic>=2.0.0
jsonschema>=4.17.0 

# Google Cloud Storage
google-cloud-storage>=2.16.0